TW-1106: Storage analytics. Refactor
