+++ src/hooks/root-hooks/use-reset-keychain-on-install.ts
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { resetKeychainOnInstallAction } from '../store/root-state.actions';
-import { useFirstAppLaunchSelector } from '../store/settings/settings-selectors';
+import { resetKeychainOnInstallAction } from 'src/store/root-state.actions';
+import { useFirstAppLaunchSelector } from 'src/store/settings/settings-selectors';
 
 export const useResetKeychainOnInstall = () => {
   const dispatch = useDispatch();
