+++ src/hooks/root-hooks/use-reset-loading.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { setIsShowLoaderAction } from '../store/settings/settings-actions';
+import { setIsShowLoaderAction } from 'src/store/settings/settings-actions';
 
 export const useResetLoading = () => {
   const dispatch = useDispatch();
