+++ src/navigator/root-stack.tsx
@@ -7,10 +7,10 @@ import React, { useCallback, useState } from 'react';
 import { useModalOptions } from 'src/components/header/use-modal-options.util';
 import { Loader } from 'src/components/loader/loader';
 import { isAndroid } from 'src/config/system';
+import { useRootHooks } from 'src/hooks/root-hooks';
 import { useAppSplash } from 'src/hooks/use-app-splash.hook';
 import { useDevicePasscode } from 'src/hooks/use-device-passcode.hook';
 import { useNetworkInfo } from 'src/hooks/use-network-info.hook';
-import { useRootHooks } from 'src/hooks/use-root-hooks';
 import { AddAssetModal } from 'src/modals/add-asset-modal/add-asset-modal';
 import { CollectibleModal } from 'src/modals/collectible-modal/collectible-modal';
 import { ConfirmationModal } from 'src/modals/confirmation-modal/confirmation-modal';
